Pushing changes on Unit base class
not abstract anymore and removed sub-classes. Will retrieve values from database in the future
