+++ Assets/Resources/prefabs/Units/ZombieFairy.prefab
@@ -10,8 +10,8 @@ GameObject:
   m_Component:
   - component: {fileID: 5258543067510660644}
   - component: {fileID: 7465976952693087701}
-  - component: {fileID: 697174012584984840}
   - component: {fileID: 6371807924774457302}
+  - component: {fileID: 6313655576437932574}
   m_Layer: 0
   m_Name: ZombieFairy
   m_TagString: Untagged
@@ -84,28 +84,6 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!114 &697174012584984840
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3923218913320148671}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 97212c6a1a1281a4f93bcbb0267e23f1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  health: 0
-  speed: 0
-  angle: 0
-  activeBullet: 
-  bezierCurveT: 0
-  controlPoints: []
-  curr_wp: 0
-  idle: 0
-  spriteRen: {fileID: 0}
-  animator: {fileID: 0}
 --- !u!95 &6371807924774457302
 Animator:
   serializedVersion: 3
@@ -125,3 +103,15 @@ Animator:
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorControllerStateOnDisable: 0
+--- !u!114 &6313655576437932574
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3923218913320148671}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 730f1f30400c7b942ad791b66077b9c8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
