--- Assets/Scripts/Rotation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0464b3f35ea34e646aba0a7248307dd7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
