+++ Assets/Scripts/SQLite Database/DatabaseHandler.cs
@@ -4,5 +4,8 @@ using UnityEngine;
 
 public static class DatabaseHandler
 {
-
+    public static void AddDatabaseEntry(string table, string type, params float[] values)
+    {
+        //TODO access database and add entry for the BulletData OR UnitData
+    }
 }
