--- Assets/Scripts/Unit/Boss.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8a4972f826061f248bbc563ac7adbafe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
