--- Assets/Scripts/Unit/Fairy.cs
@@ -1,5 +1,0 @@
-
-public class Fairy : Unit
-{
-
-}
