--- Assets/Scripts/Unit/Fairy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 97212c6a1a1281a4f93bcbb0267e23f1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
